Move AnimateEntity.h for consistency
