+++ src/st/animate_entity.h
