+++ src/st/mad/EDB8.c
@@ -285,7 +285,7 @@ void PreventEntityFromRespawning(Entity* entity) {
     }
 }
 
-#include "st/AnimateEntity.h"
+#include "../animate_entity.h"
 
 u8 func_80191F24(u8 frames[], Entity* self, u8 arg2) {
     u16 animFrameStart = self->animFrameIdx * 2;
