+++ src/st/no3/41C80.c
@@ -363,7 +363,7 @@ void PreventEntityFromRespawning(Entity* entity) {
     }
 }
 
-#include "st/AnimateEntity.h"
+#include "../animate_entity.h"
 
 u8 func_801C4E4C(u8 frames[], Entity* self, u8 arg2) {
     u16 animFrameStart = self->animFrameIdx * 2;
