+++ src/st/np3/394F0.c
@@ -350,7 +350,7 @@ void PreventEntityFromRespawning(Entity* entity) {
     }
 }
 
-#include "st/AnimateEntity.h"
+#include "../animate_entity.h"
 
 u8 func_801BC6BC(u8 frames[], Entity* self, u8 arg2) {
     u16 animFrameStart = self->animFrameIdx * 2;
