+++ src/st/nz0/39908.c
@@ -356,7 +356,7 @@ void PreventEntityFromRespawning(Entity* entity) {
     }
 }
 
-#include "st/AnimateEntity.h"
+#include "../animate_entity.h"
 
 u8 func_801BCAD4(u8 frames[], Entity* self, u8 arg2) {
     u16 animFrameStart = self->animFrameIdx * 2;
