+++ src/st/st0/31CA0.c
@@ -357,7 +357,7 @@ void PreventEntityFromRespawning(Entity* entity) {
     }
 }
 
-#include "st/AnimateEntity.h"
+#include "../animate_entity.h"
 
 u8 func_801B4AF0(u8 frames[], Entity* self, u8 arg2) {
     u16 animFrameStart = self->animFrameIdx * 2;
